moved setup created image uploading

+++ config/setup.php
@@ -1,28 +1,7 @@
 <?php
-// include('connect_database.php');
-if (isset($_POST['value']))
+include('database.php');
+if (isset($_POST['value']) && $_POST['value'] == $dbpassword)
 {
-	$dbservername = "178.128.45.163";
-	$dbuser = "notroot";
-	$dbpassword = $_POST['value'];
-	$dbname = "camagru_db";
-
-	// $conn = mysqli_connect($dbservername, $dbuser, $dbpassword);
-
-
-	$opt = [
-		PDO::ATTR_ERRMODE            => PDO::ERRMODE_EXCEPTION,
-		PDO::ATTR_DEFAULT_FETCH_MODE => PDO::FETCH_ASSOC,
-		PDO::ATTR_EMULATE_PREPARES   => false,
-	];
-	
-	
-	$conn = new PDO("mysql:host=$dbservername", $dbuser, $dbpassword);
-	$conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-
-
-		
-
 
 	$deleteDB = " DROP DATABASE $dbname";
 	
@@ -46,28 +25,19 @@ if (isset($_POST['value']))
 		echo "Database created successfully\n ".rand(0,100)."<BR /> ";
 		$conn->exec($users);
 		echo "User Table created successfully\n <BR />";
-		// if ($conn->exec($users))
-		// {
-			// echo "User Table created successfully\n <BR />";
-		// }
-		// else
-		// {
-		// 	echo "User Table FAILED\n <BR />"; 
-		// }
-		
 		
 	}
 	else
 	{
 		echo "Error creating database: " . $conn->error;
 	}
-	$conn->close();
 }
 ?>
 
 <html>
 <head>
 	<title>Database Controls</title>
+	<?php include_once("../base.php"); ?>
 	<link rel="stylesheet" type ="text/css" href="./Users/reg_style.css">
 </head>
 <body>
